+++ source/Classes/001.ChocolateyReason.ps1
@@ -1,0 +1,8 @@
+class ChocolateyReason
+{
+    [DscProperty(NotConfigurable)]
+    [string] $Code
+
+    [DscProperty(NotConfigurable)]
+    [string] $Phrase
+}