+++ source/private/Get-ChocolateyDefaultArgument.ps1
@@ -238,7 +238,7 @@ Should install arguments be used exclusively without appending to current packag
 #>
 function Get-ChocolateyDefaultArgument
 {
-    [CmdletBinding()]
+    [CmdletBinding(SupportsShouldProcess = $true, ConfirmImpact = 'None')]
     [OutputType([collections.generic.list[string]])]
     param
     (
@@ -473,7 +473,7 @@ function Get-ChocolateyDefaultArgument
     process
     {
         [collections.generic.list[string]] $ChocoArguments = [collections.generic.list[string]]::new()
-        $ChocoArguments = switch ($PSBoundParameters.Keys)
+        $Arguments = switch ($PSBoundParameters.Keys)
         {
             'Value'
             {
@@ -789,8 +789,16 @@ function Get-ChocolateyDefaultArgument
             }
         }
 
-        $null = $ChocoArguments.Add('--no-progress')
-        $null = $ChocoArguments.Add('--limit-output')
-        return $ChocoArguments
+
+        if ($PSCmdlet.ShouldProcess('Returning Choco arguments','return'))
+        {
+            $Arguments.Foreach{
+                $null = $ChocoArguments.Add($_)
+            }
+
+            $null = $ChocoArguments.Add('--no-progress')
+            $null = $ChocoArguments.Add('--limit-output')
+            return $ChocoArguments
+        }
     }
 }